+++ src/main/java/main/model/dto/LastResultColorsDto.java
@@ -1,0 +1,10 @@
+package main.model.dto;
+
+import lombok.Data;
+
+@Data
+public class LastResultColorsDto extends BaseDto {
+    private String resolution_colors;
+    private String result_colors;
+    private String result_ids;
+}